# Improve efficiency of modifying resource.dat file