Did a loose benchmark:

* Old startup time:
    * With no resource.dat.bak: around 50 seconds
    * With resource.dat.bak: around 65 seconds
* After changes
    * With no resource.dat.bak: around 27 seconds
    * With resource.dat.bak: around 30 seconds

Breakdown of improvement by change:
* Switching from list search to hash search: 3-4 seconds
* Removing stream copying: 14 seconds with no resource.dat.bak, around 30 seconds with it
* Faster signature scan: around 3 seconds

Note: startup times include the loading bar, which takes about 7 seconds. Meaning all the modloader preloading stuff takes around 40-55 seconds before changes, and around 20-23 seconds after.

----

Note: we could possibly save a few more seconds if we inject our signature into the start of `resource.dat` and break the signature scan loop as soon as its found. Based on how fast it ran with the broken signature scan code I had earlier, I doubt it would be much more than a second gain, so probably not worth it (plus, not sure if there is risk with changing the order of files)

----

Testing:
* If you contribute to the modloader, you know how to test
* If not, test by installing the mod loader using this link: [ITB-ModLoader-2.7.0-faster-resource-dat.zip](https://github.com/itb-community/ITB-ModLoader/files/9174852/ITB-ModLoader-2.7.0-faster-resource-dat.zip)