@Lemonymous 2022-07-26
I have not gone through the code, but I did test it on my system. On my system, it cuts the loading of the mod loader without mods to a third of `master`.