@KnightMiner 2022-07-26
Not sure what could cause that apart from a mod manually modifying the resource.dat, there is no path where `assertExists` would have worked with the old code where it does not work with the new code